+++ class/Renderer/Smarty3.php
@@ -34,9 +34,7 @@ class Ethna_Renderer_Smarty3 extends Ethna_Renderer
         parent::__construct($controller);
 
         // get renderer config
-        $smarty_config = isset($this->config['smarty3'])
-            ? $this->config['smarty3']
-            : array();
+        $smarty_config = $this->config;
         $this->loadEngine($smarty_config);
 
         $this->engine = new Smarty();
@@ -48,8 +46,8 @@ class Ethna_Renderer_Smarty3 extends Ethna_Renderer
         $this->setTemplateDir($template_dir);
         $this->compile_dir = $compile_dir;
 
-        $this->engine->template_dir = $template_dir;
-        $this->engine->compile_dir = $compile_dir;
+        $this->engine->setTemplateDir($template_dir);
+        $this->engine->setCompileDir($compile_dir);
         $this->engine->compile_id = md5($this->template_dir);
 
         if (isset($smarty_config['left_delimiter'])) {
@@ -60,14 +58,19 @@ class Ethna_Renderer_Smarty3 extends Ethna_Renderer
         }
 
         // make compile dir
-        if (is_dir($this->engine->compile_dir) === false) {
-            Ethna_Util::mkdir($this->engine->compile_dir, 0755);
+        if (is_dir($this->engine->getCompileDir()) === false) {
+            Ethna_Util::mkdir($this->engine->getCompileDir(), 0755);
         }
 
-        $this->engine->plugins_dir = array_merge(
+        $this->engine->setPluginsDir(array_merge(
             $controller->getDirectory('plugins'),
             array(ETHNA_BASE . '/class/Plugin/Smarty', SMARTY_DIR . 'plugins')
-        );
+        ));
+    }
+
+    public function getName()
+    {
+        return 'smarty3';
     }
 
     /**
@@ -220,6 +223,31 @@ class Ethna_Renderer_Smarty3 extends Ethna_Renderer
         // プラグインを登録する
         parent::setPlugin($name, $type, $plugin);
         $this->engine->$register_method($name, $plugin);
+    }
+
+    /**
+     * compiled template used by i18n command
+     *
+     * @return string or Ethna_Error
+     */
+    public function getCompiledContent($file)
+    {
+        $engine = $this->getEngine();
+        $tpl = $engine->createTemplate($file);
+
+        $compiled = $tpl->source->getCompiled($tpl);
+        if (!$compiled->isCompiled) {
+            $tpl->compileTemplateSource();
+        }
+
+        $compile_result = file_get_contents($compiled->filepath);
+        if (empty($compile_result)) {
+            return Ethna::raiseError(
+                "Could not compile template file : $file"
+            );
+        }
+
+        return $compile_result;
     }
 }
 // }}}