+++ test/Renderer/Smarty_Test.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ *  Renderer_Smarty_Test.php
+ */
+
+/**
+ *  Test case for Ethna_Renderer_Smarty
+ *
+ *  @access public
+ */
+class Ethna_Renderer_Smarty_Test extends Ethna_UnitTestBase
+{
+    // not impemented yet
+}