+++ test/data/template/test.php
@@ -1,0 +1,2 @@
+miyazaki: <?php echo $miyazaki ?>.
+tabe: <?php echo $tabe ?>.